Changes to be committed: 	modified:   CMakeLists.txt 	renamed:    build/.cmake/api/v1/reply/codemodel-v2-e258f46de39a62cb04f4.json -> build/.cmake/api/v1/reply/codemodel-v2-5bb24fcc45a8855b19f2.json 	renamed:    build/.cmake/api/v1/reply/directory-.-Debug-1f306fce3a1e52d0a9f0.json -> build/.cmake/api/v1/reply/directory-.-Debug-ed42700fb7379a5dabcc.json 	renamed:    build/.cmake/api/v1/reply/index-2024-05-25T08-23-11-0232.json -> build/.cmake/api/v1/reply/index-2024-05-25T08-35-21-0079.json 	renamed:    build/.cmake/api/v1/reply/target-clib_shared-Debug-ba28b69d6a0c97f55f9c.json -> build/.cmake/api/v1/reply/target-clib_shared-Debug-3cf7b07628bb3bac47b3.json 	renamed:    build/.cmake/api/v1/reply/target-clib_static-Debug-8262733e98a7867aecda.json -> build/.cmake/api/v1/reply/target-clib_static-Debug-7e240ee05ab0e2a1e02f.json 	modified:   build/.ninja_deps 	modified:   build/.ninja_log 	modified:   build/CMakeFiles/clib_shared.dir/clib.c.o 	modified:   build/CMakeFiles/clib_static.dir/clib.c.o 	modified:   build/cmake_install.cmake 	modified:   build/install_manifest.txt 	modified:   build/libclib.a 	modified:   build/libclib.so 	modified:   clib.c 	modified:   clib.h

+++ clib.c
@@ -1,5 +1,6 @@
 #include "clib.h"
 
+#include "def.h"
 #include "errno.h"
 #include "pwd.h"
 #include "stdio.h"
@@ -266,8 +267,15 @@ void clear_buffer(char* const buffer, size_t const size)
 }
 
 
-void rmdir_asm(const char* const path)
+u8 clib_rmdir(const char* const path)
 {
+    if (path == NULL)
+    {
+        errno = EINVAL;
+        perror("clib_rmdir");
+        return NULL_FAILURE;
+    }
+
     long result;
     asm volatile("movq %1, %%rdi\n"  // Move the path argument to the first argument register (rdi)
                  "movl $84, %%eax\n" // Move the syscall number for rmdir to eax
@@ -281,7 +289,11 @@ void rmdir_asm(const char* const path)
     if (result < 0)
     {
         errno = -result;
+        perror("clib_rmdir");
+        return FAILURE;
     }
+
+    return SUCCESS;
 }
 
 
@@ -395,7 +407,7 @@ u8 append_str(char* dest, const char* src, size_t dest_size)
     return SUCCESS;
 }
 
-void* allocate(size_t size)
+void* allocate(size_t const size)
 {
     void* result;
     long  mmap_result;
